⬆️ Upgrade to Next.js 16
Follows https://nextjs.org/docs/app/guides/upgrading/version-16.

- next 15.5.18 → 16.2.12, react/react-dom 19.2.3 → 19.2.8,
  @next/third-parties and @next/bundle-analyzer → 16.2.12
- Rename middleware.ts → proxy.ts and its export to `proxy`; move the
  noIndex helper to lib/ so nothing sits next to the new convention
- Drop the webpack block from next.config.mjs so `next build` runs on
  Turbopack: the @svgr/webpack rule had no callers (every SVG is
  referenced by public path) and the @tinacms/mdx alias is already
  mirrored in `turbopack.resolveAlias`
- Replace next-plugin-preval (webpack-only, peer-capped at Next 15) with
  lib/platform.js, imported by both tailwind.config.js and ColorPalette
- Remove `eslint.ignoreDuringBuilds` — the option is gone in 16 and
  `next build` no longer lints
- Drop `--turbopack` from the dev script; it is the default now
