--- components/util/preval/get-tailwind-platform-colors.preval.ts
@@ -1,9 +1,0 @@
-import preval from "next-plugin-preval";
-
-import { platform } from "../../../tailwind.config.js";
-
-async function getTailwindPlatformColors() {
-  return platform.map((p) => ({ ...p, className: `bg-platform-${p.name}` }));
-}
-
-export default preval(getTailwindPlatformColors());
