+++ package.json
@@ -15,7 +15,7 @@
   },
   "scripts": {
     "preinstall": "npx only-allow pnpm",
-    "dev": "tinacms dev -c \"next dev --turbopack\"",
+    "dev": "tinacms dev -c \"next dev\"",
     "build": "tinacms build --content=local && next build && next-sitemap --config next-sitemap.config.js",
     "partytown": "partytown copylib public/~partytown",
     "postbuild": "next-sitemap --config next-sitemap.config.js",
@@ -33,7 +33,6 @@
     "@jest/globals": "^30.2.0",
     "@next/eslint-plugin-next": "15.5.1",
     "@playwright/test": "^1.56.1",
-    "@svgr/webpack": "^8.1.0",
     "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/react": "^16.3.1",
     "@tinacms/cli": "2.5.3",
@@ -67,8 +66,8 @@
     "@headlessui/tailwindcss": "^0.2.2",
     "@microsoft/applicationinsights-react-js": "^19.3.6",
     "@microsoft/applicationinsights-web": "^3.3.8",
-    "@next/bundle-analyzer": "16.0.3",
-    "@next/third-parties": "15.5.6",
+    "@next/bundle-analyzer": "16.2.12",
+    "@next/third-parties": "16.2.12",
     "@radix-ui/react-accordion": "^1.2.12",
     "@radix-ui/react-checkbox": "^1.3.3",
     "@radix-ui/react-dropdown-menu": "^2.1.16",
@@ -97,14 +96,13 @@
     "lodash": "^4.18.1",
     "lru-cache": "^11.1.0",
     "lucide-react": "^0.545.0",
-    "next": "15.5.18",
-    "next-plugin-preval": "^1.2.8",
+    "next": "16.2.12",
     "next-seo": "^6.8.0",
     "nextjs-breadcrumbs2": "^2.0.4",
     "node-cache": "^5.1.2",
-    "react": "19.2.3",
+    "react": "19.2.8",
     "react-device-detect": "^2.2.3",
-    "react-dom": "19.2.3",
+    "react-dom": "19.2.8",
     "react-google-recaptcha": "^3.1.0",
     "react-icons": "^5.5.0",
     "react-jotform": "^1.0.2",
