+++ tailwind.config.js
@@ -3,53 +3,7 @@ import typographyPlugin from "@tailwindcss/typography";
 import colors from "tailwindcss/colors";
 import plugin from "tailwindcss/plugin";
 
-export const platform = [
-  {
-    name: "angular",
-    text: "Angular",
-    color: "#DD0031",
-  },
-  {
-    name: "dotnet",
-    text: ".NET",
-    color: "#5C2D91",
-  },
-  {
-    name: "visualstudio",
-    text: "Visual Studio",
-    color: "#9455CE",
-  },
-  {
-    name: "blazor",
-    text: "Blazor",
-    color: "#5C2D91",
-  },
-  {
-    name: "xamarin",
-    text: "Xamarin",
-    color: "#3498DB",
-  },
-  {
-    name: "azure",
-    text: "Azure",
-    color: "#0088D5",
-  },
-  {
-    name: "sharepoint",
-    text: "SharePoint",
-    color: "#038185",
-  },
-  {
-    name: "powerbi",
-    text: "PowerBI",
-    color: "#F2C811",
-  },
-  {
-    name: "tina",
-    text: "TinaCMS",
-    color: "#EC4815",
-  },
-];
+import { platform } from "./lib/platform.js";
 
 // eslint-disable-next-line import/no-anonymous-default-export
 export default {
